Removed broken and buggy software.
The software being removed is notorious buggy and hard to use. It adds complexity to any environment it is installed in and usually breaks them randomly, in a spectacular way.

The software is made up from lies and broken dreams, it is considered harmful and should not be used by anyone.